// src/consumer/events/order.paid.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type OrderPaid = SeatlyEventEnvelope & {
  event_type?: "order.paid";
  version?: 1;
  payload?: {
    order_id: string;
    event_id: number;
    seat_ids: number[];
    buyer_id: number;
  };
  [k: string]: unknown;
};

export interface SeatlyEventEnvelope {
  event_id: string;
  event_type: "event.published" | "order.paid" | "order.payment_failed" | "tickets.issued";
  occurred_at: string;
  version: number;
  payload: {
    [k: string]: unknown;
  };
}